Refactor Akamai stream connector with SDK config and metadata
