+++ stream/akamai/__metadata__/connector_manifest.json
@@ -1,0 +1,22 @@
+{
+  "title": "Akamai Stream Connector",
+  "slug": "akamai-stream",
+  "description": "The Akamai Stream connector synchronizes IPv4 STIX Indicators from OpenCTI to an Akamai Client List. It listens to the OpenCTI live stream and automatically adds or removes IP indicators in the configured Akamai Client List, enabling near real-time enforcement of threat intelligence data.",
+  "short_description": "Synchronizes OpenCTI IPv4 indicators with an Akamai Client List via stream.",
+  "logo": "stream/akamai-stream/__metadata__/logo.png",
+  "use_cases": [
+    "Threat Intelligence Distribution",
+    "Network Protection"
+  ],
+  "verified": false,
+  "last_verified_date": null,
+  "playbook_supported": false,
+  "max_confidence_level": 80,
+  "support_version": ">= 6.7.0",
+  "subscription_link": "https://www.akamai.com/",
+  "source_code": "https://github.com/OpenCTI-Platform/connectors/tree/master/stream/akamai-stream",
+  "manager_supported": false,
+  "container_version": "latest",
+  "container_image": "opencti/connector-akamai-stream",
+  "container_type": "STREAM"
+}
