+++ stream/akamai/src/akamai_connector/__init__.py
@@ -1,0 +1,4 @@
+# Expose AkamaiConnector class
+from .connector import AkamaiConnector
+
+__all__ = ["AkamaiConnector"]
