+++ stream/akamai/src/config.yml.sample
@@ -1,8 +1,10 @@
+# --- OpenCTI configuration ---
 opencti:
-  url: https://your-opencti-instance
+  url: http://localhost
   token: CHANGE_ME
 
 
+# --- Akamai configuration ---
 akamai:
   base_url: https://your-host.akamaiapis.net
   client_token: CHANGE_ME
@@ -10,4 +12,7 @@ akamai:
   access_token: CHANGE_ME
   client_list_id: CHANGE_ME
 
-ssl_verify: false
+
+ssl_verify: true
+
+
