+++ stream/akamai/src/main.py
@@ -1,32 +1,39 @@
-import os
 import sys
 import time
 import traceback
 
 from pycti import OpenCTIConnectorHelper
-from stream_connector.connector import AkamaiConnector
+from akamai_connector.connector import AkamaiConnector
+from settings import get_config
+
 
 if __name__ == "__main__":
     try:
         # Initialize OpenCTI helper
-        # It automatically reads OPENCTI_* variables from environment
+        # This automatically loads OpenCTI-related environment variables
+
         helper = OpenCTIConnectorHelper({})
 
-        # Instantiate connector using environment variables
+        # Load connector configuration
+        # Configuration is centralized in settings.py for maintainability
+        config = get_config(helper)
+
+        # Instantiate the Akamai connector with the loaded configuration
+        # Using **config allows flexible and clean parameter passing
         connector = AkamaiConnector(
             helper=helper,
-            base_url=os.environ["AKAMAI_BASE_URL"],
-            client_token=os.environ["AKAMAI_CLIENT_TOKEN"],
-            client_secret=os.environ["AKAMAI_CLIENT_SECRET"],
-            access_token=os.environ["AKAMAI_ACCESS_TOKEN"],
-            client_list_id=os.environ["AKAMAI_CLIENT_LIST_ID"],
+            **config
         )
 
-        # Start listening to OpenCTI live stream
-        connector.start()
+        # Start the connector
+        # This will listen to the OpenCTI live stream and process events
+        connector.run()
 
     except Exception:
-        # Print full stack trace in case of crash
+        
         traceback.print_exc()
+
+        # Small delay before exit to avoid crash loops in container environments
         time.sleep(10)
+
         sys.exit(1)
