--- stream/akamai/src/stream_connector/__init__.py
@@ -1,4 +1,0 @@
-# Expose AkamaiConnector class
-from .connector import AkamaiConnector
-
-__all__ = ["AkamaiConnector"]
